bebob-runtime: apogee: use protocol implementation of convert parameters
